+++ CHANGELOG.md
@@ -16,7 +16,10 @@ Possible log types:
 - [added] The `Sensors`, `PeopleNowPresentSensor` and `TemperatureSensor`
   structs now derive `Default` (#84)
 - [added] Basic support for the new v14 API (#85) This is a breaking change since
-  it changes the `api` field of `Status` to `Option<String>`.
+  it changes the `api` field of `Status` to `Option<String>`. Includes
+  * Add `xmpp` field in `Contact` struct and deprecate the `jabber` field
+  * Add `xmpp` field to `Keymaster`
+
 
 ### v0.7.0 (2019-08-22)